refactor: reducers separados
