+++ client/src/store/index.js
@@ -1,9 +1,10 @@
 import {configureStore} from '@reduxjs/toolkit';
 import music from "./slice/index"
-
+import user from "./slice/user.js";
 
 export default configureStore({
-    reducer: {
-        music
-    }
+  reducer: {
+    music,
+    user,
+  }
 })
